backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
pymongo==4.8.0
python-dotenv==1.0.1
pycryptodome==3.20.0
PyJWT==2.9.0
cryptography==43.0.0
httpx==0.27.0
pydantic==2.9.0
